Split code and make structure more consistent
