+++ cuba_weather/__init__.py
@@ -1,3 +1,5 @@
+from .api import RCApiClient, URL
+from .console import main
 from .finder import distance, get_location, get_suggestions
 from .locations import locations
-from .weather import API, main, RCApiClient
+from .weather import Weather
